Remove EnvironmentPlugin usage for DefinePlugin
