+++ lib/config-generator.js
@@ -32,7 +32,6 @@ const versioningPluginUtil = require('./plugins/versioning');
 const variableProviderPluginUtil = require('./plugins/variable-provider');
 const cleanPluginUtil = require('./plugins/clean');
 const definePluginUtil = require('./plugins/define');
-const environmentPluginUtil = require('./plugins/environment');
 const terserPluginUtil = require('./plugins/terser');
 const optimizeCssAssetsUtil = require('./plugins/optimize-css-assets');
 const vuePluginUtil = require('./plugins/vue');
@@ -425,10 +424,6 @@ class ConfigGenerator {
         cleanPluginUtil(plugins, this.webpackConfig);
 
         definePluginUtil(plugins, this.webpackConfig);
-
-        if (this.webpackConfig.shouldInjectEnvironmentVariables) {
-            environmentPluginUtil(plugins, this.webpackConfig);
-        }
 
         notifierPluginUtil(plugins, this.webpackConfig);
