+++ lib/plugins/define.js
@@ -9,10 +9,12 @@
 
 'use strict';
 
+const path = require('path');
 const webpack = require('webpack');
 const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const PluginPriorities = require('./plugin-priorities');
 const applyOptionsCallback = require('../utils/apply-options-callback');
+const { loadEnv } = require('../env-helper');
 
 /**
  * @param {Array} plugins
@@ -32,4 +34,23 @@ module.exports = function(plugins, webpackConfig) {
         ),
         priority: PluginPriorities.DefinePlugin
     });
+
+    if (webpackConfig.shouldInjectEnvironmentVariables) {
+        const envVars = Object.assign({}, process.env);
+        const envVarsPath = path.isAbsolute(webpackConfig.environmentVariablesPath)
+            ? webpackConfig.environmentVariablesPath
+            : path.resolve(webpackConfig.getContext(), webpackConfig.environmentVariablesPath);
+
+        loadEnv(envVars, envVarsPath, webpackConfig);
+
+        const defs = Object.keys(envVars).reduce((acc, key) => {
+            acc[`process.env.${key}`] = JSON.stringify(envVars[key]);
+            return acc;
+        }, {});
+
+        plugins.push({
+            plugin: new webpack.DefinePlugin(defs),
+            priority: PluginPriorities.DefinePlugin
+        });
+    }
 };
