+++ lib/plugins/plugin-priorities.js
@@ -18,7 +18,6 @@ module.exports = {
     ProvidePlugin: 90,
     CleanWebpackPlugin: 80,
     DefinePlugin: 70,
-    EnvironmentPlugin: 65,
     WebpackNotifier: 60,
     VueLoaderPlugin: 50,
     FriendlyErrorsWebpackPlugin: 40,
